+++ src/com/kaylerrenslow/armaDialogCreator/gui/main/treeview/ControlGroupTreeItemEntry.java
@@ -18,6 +18,7 @@ public class ControlGroupTreeItemEntry extends ControlTreeItemEntry {
 		this.controlGroup = group;
 	}
 
+	@NotNull
 	public ArmaControlGroup getControlGroup() {
 		return controlGroup;
 	}