+++ CHANGES.md
@@ -2,10 +2,11 @@
 * 
 
 **Changed:**
-* optimized EditorComponentTreeView
+* 
 
 **Fixed:**
-* 
+* ProjectExporter wasn't writing control group classes in right spot
+* EditorComponentTreeView not handling control groups properly
 
 **Notes:**
 * May 31: we should have default value providers in one xml file for every control.